refactor(who_vax): moving unvax population calculation to garden
