+++ etl/steps/data/garden/who/2022-07-17/who_vaccination.py
@@ -1,7 +1,9 @@
 import json
+from functools import reduce
 from typing import List, cast
 
 import pandas as pd
+from owid import catalog
 from owid.catalog import Dataset, Table
 from owid.catalog.utils import underscore_table
 from owid.datautils import geo
@@ -10,14 +12,13 @@ from structlog import get_logger
 from etl.helpers import Names
 from etl.paths import DATA_DIR
 
+N = Names(__file__)
+
+UNWPP = DATA_DIR / "garden/un/2022-07-11/un_wpp"
+
 log = get_logger()
 
-##
-# To Do:
-# - Add in population of target population to calculate unvaccinated population. At least for one year olds.
-# naming conventions
 N = Names(__file__)
-N = Names("/Users/fionaspooner/Documents/OWID/repos/etl/etl/steps/data/garden/who/2022-07-17/who_vaccination.py")
 
 
 def run(dest_dir: str) -> None:
@@ -38,26 +39,15 @@ def run(dest_dir: str) -> None:
     ds_garden = Dataset.create_empty(dest_dir)
     ds_garden.metadata = ds_meadow.metadata
 
-    # may need to combine japanese encephalitis and japanese encephalitis first dose
-    # We use only the official figures
-    df = df[df["coverage_category"] == "OFFICIAL"]
-    df = df.dropna(subset="coverage")
-    df = df.drop(columns=["index", "group", "antigen", "coverage_category", "coverage_category_description"])
-    df = df.pivot_table(
-        values=["coverage", "doses", "target_number"],
-        columns=["antigen_description"],
-        index=[
-            "country",
-            "year",
-        ],
+    tb_garden = clean_and_format_data(df)
+
+    pop_one_yr = get_population_one_year_olds()
+    pop_one_yr = pop_one_yr.rename(columns={"location": "country", "value": "population"}).drop(
+        columns=["metric", "variant", "sex", "age"]
     )
-    df.columns = df.columns.to_series().apply("_".join).str.replace("coverage_", "")
-    df = df.reset_index()
+    # table = N.garden_dataset["who_vaccination"]
+    tb_garden = calculate_vaccinated_unvaccinated_population(tb_garden, pop_one_yr)
 
-    tb_garden = underscore_table(Table(df))
-    cols = tb_garden.drop(["country", "year"], axis=1).columns
-
-    tb_garden.loc[:, cols] = tb_garden.loc[:, cols].astype(float).round(2)
     tb_garden.metadata = tb_meadow.metadata
     ds_garden.metadata.update_from_yaml(N.metadata_path)
     tb_garden.update_metadata_from_yaml(N.metadata_path, "who_vaccination")
@@ -94,3 +84,75 @@ def harmonize_countries(df: pd.DataFrame) -> pd.DataFrame:
         )
 
     return df
+
+
+def get_population_one_year_olds() -> pd.DataFrame:
+    un_wpp_data = catalog.Dataset(UNWPP)
+    pop = un_wpp_data["population"].reset_index()
+    pop_one_yr = pop[
+        (pop["age"] == "1") & (pop["variant"] == "estimates") & (pop["metric"] == "population") & (pop["sex"] == "all")
+    ]
+    pop_one_yr = pd.DataFrame(pop_one_yr)
+    return pop_one_yr
+
+
+def calculate_vaccinated_unvaccinated_population(table: Table, pop_one_yr: pd.DataFrame) -> pd.DataFrame:
+    # vaccines where the coverage is measured as % of one-year olds
+    vax_one_year_olds = [
+        "bcg",
+        "dtp_containing_vaccine__1st_dose",
+        "dtp_containing_vaccine__3rd_dose",
+        "hepb3",
+        "hepb__birth_dose__given_within_24_hours_of_birth",
+        "hib3",
+        "inactivated_polio_containing_vaccine__1st_dose",
+        "measles_containing_vaccine__1st_dose",
+        "measles_containing_vaccine__2nd_dose",
+        "pneumococcal_conjugate_vaccine__final_dose",
+        "polio__3rd_dose",
+        "rubella_containing_vaccine__1st_dose",
+        "rotavirus__last_dose",
+        "yellow_fever_vaccine",
+    ]
+    cov_pop = table[["country", "year"] + vax_one_year_olds].merge(pop_one_yr, on=["country", "year"])
+    vax_pop = cov_pop
+    vax_pop[vax_one_year_olds] = (
+        vax_pop[vax_one_year_olds].multiply(0.01).multiply(vax_pop["population"], axis="index").round(0).astype("Int64")
+    )
+    unvax_pop = vax_pop
+    vax_pop = vax_pop.rename(columns={c: c + "_vaccinated" for c in vax_pop.columns if c in vax_one_year_olds})
+    vax_pop = vax_pop.drop(columns=["population"])
+
+    unvax_pop[vax_one_year_olds] = (
+        unvax_pop[vax_one_year_olds].sub(unvax_pop["population"], axis="index").multiply(-1).astype("Int64")
+    )
+    unvax_pop = unvax_pop.rename(columns={c: c + "_unvaccinated" for c in unvax_pop.columns if c in vax_one_year_olds})
+    unvax_pop = unvax_pop.drop(columns=["population"])
+    data_frames = [table, vax_pop, unvax_pop]
+
+    df_merged = reduce(lambda left, right: pd.merge(left, right, on=["country", "year"], how="outer"), data_frames)
+    return df_merged
+
+
+def clean_and_format_data(df: pd.DataFrame) -> pd.DataFrame:
+    # may need to combine japanese encephalitis and japanese encephalitis first dose
+    # We use only the official figures
+    df = df[df["coverage_category"] == "OFFICIAL"]
+    df = df.dropna(subset="coverage")
+    df = df.drop(columns=["index", "group", "antigen", "coverage_category", "coverage_category_description"])
+    df = df.pivot_table(
+        values=["coverage", "doses", "target_number"],
+        columns=["antigen_description"],
+        index=[
+            "country",
+            "year",
+        ],
+    )
+    df.columns = df.columns.to_series().apply("_".join).str.replace("coverage_", "")
+    df = df.reset_index()
+
+    tb_garden = underscore_table(Table(df))
+    cols = tb_garden.drop(["country", "year"], axis=1).columns
+
+    tb_garden.loc[:, cols] = tb_garden.loc[:, cols].astype(float).round(2)
+    return tb_garden
